Move package manager stuff to other file location
