+++ src/core/userDialog/PerformUserDialog.test.ts
@@ -1,8 +1,8 @@
 import * as PerformUserDialog from './PerformUserDialog'
 import { performUserDialog, promptMetadata } from './PerformUserDialog'
 import * as SelectExtensions from './SelectExtensions'
-import * as ChoosePackageManager from './ChoosePackageManager'
-import { PackageManager } from './ChoosePackageManager'
+import * as ChoosePackageManager from '../PackageManagers'
+import { PackageManager } from '../PackageManagers'
 import { filter, Subject } from 'rxjs'
 import { Extension } from '../Extension'
 import { Answers, DistinctQuestion, ListQuestion } from 'inquirer'
